services/embedding_service/requirements.txt
functions-framework
google-cloud-storage
google-cloud-firestore
google-genai
PyPDF2
asyncio
aiohttp
